+++ CHANGELOG.md
@@ -6,6 +6,15 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
+## [1.3.0] - 2016-12-13
+### Added
+
+* Regular springs
+
+### Changed
+
+* String representation of rotary springs now shows the actual angle of the spring
+
 ## [1.2] - 2016-12-13
 ### Added